src/core/engine.py
"""
AI Teams Core Engine
核心引擎 - 协调所有agent的心跳和任务执行
"""

import asyncio
import json
import logging
from datetime import datetime
from pathlib import Path
from typing import Dict, List, Optional
import yaml


class AITeamsEngine:
    """AI Teams核心引擎"""
    
    def __init__(self, config_path: str = "config/agents.yaml"):
        self.config = self._load_config(config_path)
        self.agents = {}
        self.task_queue = []
        self.knowledge_base = {}
        self.logger = self._setup_logger()
        
    def _load_config(self, path: str) -> dict:
        """加载配置文件"""
        config_file = Path(path)
        if config_file.exists():
            with open(config_file, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        return {"agents": []}
    
    def _setup_logger(self) -> logging.Logger:
        """设置日志"""
        logger = logging.getLogger("AITeams")
        logger.setLevel(logging.INFO)
        
        # 创建logs目录
        log_dir = Path("logs")
        log_dir.mkdir(exist_ok=True)
        
        # 文件处理器
        fh = logging.FileHandler(
            log_dir / f"ai-teams-{datetime.now().strftime('%Y%m%d')}.log"
        )
        fh.setLevel(logging.INFO)
        
        # 格式
        formatter = logging.Formatter(
            '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
        fh.setFormatter(formatter)
        logger.addHandler(fh)
        
        return logger
    
    async def heartbeat(self):
        """
        心跳触发器 - 每30分钟自动执行
        这是核心：让AI主动工作，无需人工触发
        """
        self.logger.info("💓 Heartbeat triggered")
        
        # 1. 检查任务队列
        tasks = self._check_task_queue()
        
        # 2. 分析当前状态
        status = self._analyze_status()
        
        # 3. 决策下一步行动
        actions = self._decide_actions(tasks, status)
        
        # 4. 执行或分配任务
        for action in actions:
            await self._execute_action(action)
        
        # 5. 更新知识库
        self._update_knowledge()
        
        # 6. 汇报进度（如有重大进展）
        self._report_progress()
    
    def _check_task_queue(self) -> List[dict]:
        """检查任务队列"""
        queue_file = Path("data/tasks/queue.json")
        if queue_file.exists():
            with open(queue_file, 'r', encoding='utf-8') as f:
                return json.load(f)
        return []
    
    def _analyze_status(self) -> dict:
        """分析当前状态"""
        return {
            "active_tasks": len([t for t in self.task_queue if t['status'] == 'in_progress']),
            "pending_tasks": len([t for t in self.task_queue if t['status'] == 'pending']),
            "completed_today": 0,  # TODO: 实现
            "agents_online": len(self.agents),
            "last_heartbeat": datetime.now().isoformat()
        }
    
    def _decide_actions(self, tasks: List[dict], status: dict) -> List[dict]:
        """
        决策下一步行动
        这是AI"思考"的核心
        """
        actions = []
        
        # 优先级1: 处理阻塞任务
        blocked = [t for t in tasks if t['status'] == 'blocked']
        if blocked:
            actions.append({
                'type': 'resolve_blockers',
                'tasks': blocked
            })
        
        # 优先级2: 继续进行中的任务
        in_progress = [t for t in tasks if t['status'] == 'in_progress']
        if in_progress:
            actions.append({
                'type': 'continue_tasks',
                'tasks': in_progress[:3]  # 最多并行3个
            })
        
        # 优先级3: 认领新任务
        pending = [t for t in tasks if t['status'] == 'pending']
        if pending and status['active_tasks'] < 5:
            actions.append({
                'type': 'claim_tasks',
                'tasks': pending[:2]  # 每次最多认领2个
            })
        
        # 优先级4: 主动思考新机会
        if not actions or status['completed_today'] > 3:
            actions.append({
                'type': 'brainstorm',
                'focus': 'market_opportunities'
            })
        
        return actions
    
    async def _execute_action(self, action: dict):
        """执行动作"""
        action_type = action['type']
        
        if action_type == 'continue_tasks':
            for task in action['tasks']:
                await self._work_on_task(task)
        
        elif action_type == 'claim_tasks':
            for task in action['tasks']:
                await self._claim_task(task)
        
        elif action_type == 'brainstorm':
            await self._brainstorm(action['focus'])
    
    async def _work_on_task(self, task: dict):
        """处理任务"""
        self.logger.info(f"🔄 Working on task: {task['id']}")
        # TODO: 实际执行逻辑
        # 这里会调用agent来执行具体任务
    
    async def _claim_task(self, task: dict):
        """认领任务"""
        self.logger.info(f"✋ Claimed task: {task['id']}")
        task['status'] = 'in_progress'
        task['claimed_at'] = datetime.now().isoformat()
        # TODO: 保存到队列
    
    async def _brainstorm(self, focus: str):
        """头脑风暴 - 主动思考"""
        self.logger.info(f"💡 Brainstorming: {focus}")
        # TODO: 调用AI生成新想法
    
    def _update_knowledge(self):
        """更新知识库"""
        # TODO: 实现学习逻辑
        pass
    
    def _report_progress(self):
        """汇报进度"""
        # TODO: 发送消息到飞书等平台
        pass
    
    def start(self):
        """启动引擎"""
        self.logger.info("🚀 AI Teams Engine started")
        # TODO: 启动心跳循环


if __name__ == "__main__":
    engine = AITeamsEngine()
    engine.start()
