requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
apscheduler>=3.10
aiofiles>=23.0
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
